angular-google-analytics improvements (#12226)
* removes useless separated angular-google-analytics-service.d.ts file and adds this service in index.d.ts; edits namespace usage (like in angular.ui.router and other packages)

* adds reference on angular
